fix for upload github
