+++ app/src/main/java/com/glancemap/glancemapwearos/presentation/features/routetools/GpxRouteEditorOutputs.kt
@@ -1,0 +1,405 @@
+@file:Suppress(
+    "CyclomaticComplexMethod",
+    "LongMethod",
+    "LongParameterList",
+)
+
+package com.glancemap.glancemapwearos.presentation.features.routetools
+
+import com.glancemap.glancemapwearos.core.routing.RouteGeometryPoint
+import com.glancemap.glancemapwearos.presentation.features.gpx.GpxTrackDetails
+import com.glancemap.glancemapwearos.presentation.features.gpx.TrackPoint
+import com.glancemap.glancemapwearos.presentation.features.gpx.TrackPosition
+import com.glancemap.glancemapwearos.presentation.features.gpx.TrackProfile
+import com.glancemap.glancemapwearos.presentation.features.gpx.totalAscent
+import com.glancemap.glancemapwearos.presentation.features.gpx.totalDistance
+import org.mapsforge.core.model.LatLong
+
+internal fun buildRouteToolEditOutput(
+    sourcePath: String,
+    sourceFileName: String,
+    sourceTitle: String?,
+    profile: TrackProfile,
+    session: RouteToolSession,
+): RouteToolEditOutput {
+    require(session.options.toolKind == RouteToolKind.MODIFY) {
+        "Only modify actions are supported by the GPX editor."
+    }
+
+    val sourceTrack =
+        GpxTrackDetails(
+            id = sourcePath,
+            points = profile.points.map { it.latLong },
+            title = sourceTitle,
+            distance = profile.totalDistance,
+            elevationGain = profile.totalAscent,
+            startPoint = profile.points.firstOrNull()?.latLong,
+            endPoint = profile.points.lastOrNull()?.latLong,
+        )
+
+    val editedPoints =
+        when (session.options.modifyMode) {
+            RouteModifyMode.RESHAPE_ROUTE -> {
+                error("Use buildRouteToolReshapeOutput for reshape edits.")
+            }
+
+            RouteModifyMode.REPLACE_SECTION_A_TO_B -> {
+                error("Use buildRouteToolReplaceSectionOutput for routed section replacement.")
+            }
+
+            RouteModifyMode.KEEP_ONLY_A_TO_B -> {
+                val a = session.pointA ?: error("Point A is required.")
+                val b = session.pointB ?: error("Point B is required.")
+                val posA = snapToTrackPosition(target = a, sourceTrack = sourceTrack, profile = profile)
+                val posB = snapToTrackPosition(target = b, sourceTrack = sourceTrack, profile = profile)
+                slicePointsBetween(profile.points, posA, posB)
+            }
+
+            RouteModifyMode.TRIM_START_TO_HERE -> {
+                val a = session.pointA ?: error("Point A is required.")
+                val posA = snapToTrackPosition(target = a, sourceTrack = sourceTrack, profile = profile)
+                trimTrackStart(profile.points, posA)
+            }
+
+            RouteModifyMode.TRIM_END_FROM_HERE -> {
+                val b = session.pointB ?: error("Point B is required.")
+                val posB = snapToTrackPosition(target = b, sourceTrack = sourceTrack, profile = profile)
+                trimTrackEnd(profile.points, posB)
+            }
+
+            RouteModifyMode.REVERSE_GPX -> {
+                profile.points.reversed()
+            }
+        }
+
+    require(editedPoints.size >= 2) {
+        "The selected edit would produce an empty GPX."
+    }
+
+    val titleBase =
+        sourceTitle?.takeIf { it.isNotBlank() }
+            ?: sourceFileName.removeSuffix(".gpx")
+    val fileName =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> sourceFileName
+            RouteSaveBehavior.SAVE_AS_NEW -> {
+                buildEditedFileName(
+                    sourceFileName = sourceFileName,
+                    mode = session.options.modifyMode,
+                )
+            }
+        }
+
+    val title =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> titleBase
+            RouteSaveBehavior.SAVE_AS_NEW -> "$titleBase (edited)"
+        }
+
+    return RouteToolEditOutput(
+        fileName = fileName,
+        title = title,
+        points = editedPoints,
+    )
+}
+
+internal fun buildRouteToolEndpointChangeOutput(
+    sourceFileName: String,
+    sourceTitle: String?,
+    profile: TrackProfile,
+    session: RouteToolSession,
+    snappedPosition: TrackPosition,
+    routedPoints: List<RouteGeometryPoint>,
+): RouteToolEditOutput {
+    require(session.options.toolKind == RouteToolKind.MODIFY) {
+        "Only modify actions are supported by the GPX editor."
+    }
+    require(
+        session.options.modifyMode == RouteModifyMode.TRIM_START_TO_HERE ||
+            session.options.modifyMode == RouteModifyMode.TRIM_END_FROM_HERE,
+    ) {
+        "Only Change start/end are supported here."
+    }
+    require(routedPoints.size >= 2) {
+        "The routed endpoint change is empty."
+    }
+
+    val merged = ArrayList<TrackPoint>(profile.points.size + routedPoints.size)
+    when (session.options.modifyMode) {
+        RouteModifyMode.TRIM_START_TO_HERE -> {
+            routedPoints.forEach { point ->
+                appendUnique(merged, point.toTrackPoint())
+            }
+            trimTrackStart(profile.points, snappedPosition).forEach { point ->
+                appendUnique(merged, point)
+            }
+        }
+
+        RouteModifyMode.TRIM_END_FROM_HERE -> {
+            trimTrackEnd(profile.points, snappedPosition).forEach { point ->
+                appendUnique(merged, point)
+            }
+            routedPoints.forEach { point ->
+                appendUnique(merged, point.toTrackPoint())
+            }
+        }
+    }
+
+    require(merged.size >= 2) {
+        "The selected edit would produce an empty GPX."
+    }
+
+    val titleBase =
+        sourceTitle?.takeIf { it.isNotBlank() }
+            ?: sourceFileName.removeSuffix(".gpx")
+    val fileName =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> sourceFileName
+            RouteSaveBehavior.SAVE_AS_NEW -> {
+                buildEditedFileName(
+                    sourceFileName = sourceFileName,
+                    mode = session.options.modifyMode,
+                )
+            }
+        }
+    val title =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> titleBase
+            RouteSaveBehavior.SAVE_AS_NEW -> "$titleBase (edited)"
+        }
+
+    return RouteToolEditOutput(
+        fileName = fileName,
+        title = title,
+        points = merged,
+    )
+}
+
+internal fun buildRouteToolExtensionOutput(
+    sourceFileName: String,
+    sourceTitle: String?,
+    profile: TrackProfile,
+    routedPoints: List<RouteGeometryPoint>,
+): RouteToolEditOutput {
+    require(profile.points.size >= 2) {
+        "The active GPX does not contain enough points to extend."
+    }
+    require(routedPoints.size >= 2) {
+        "The routed extension is empty."
+    }
+
+    val merged = ArrayList<TrackPoint>(profile.points.size + routedPoints.size)
+    profile.points.forEach { point ->
+        appendUnique(merged, point)
+    }
+    routedPoints.forEach { point ->
+        appendUnique(merged, point.toTrackPoint())
+    }
+
+    require(merged.size >= 2) {
+        "The routed extension would produce an empty GPX."
+    }
+
+    val titleBase =
+        sourceTitle?.takeIf { it.isNotBlank() }
+            ?: sourceFileName.removeSuffix(".gpx")
+
+    return RouteToolEditOutput(
+        fileName =
+            buildEditedFileName(
+                sourceFileName = sourceFileName,
+                modeSlug = "extend",
+            ),
+        title = "$titleBase (extended)",
+        points = merged,
+    )
+}
+
+internal fun buildRouteToolReplaceSectionOutput(
+    sourcePath: String,
+    sourceFileName: String,
+    sourceTitle: String?,
+    profile: TrackProfile,
+    session: RouteToolSession,
+    routedPoints: List<RouteGeometryPoint>,
+): RouteToolEditOutput {
+    require(session.options.modifyMode == RouteModifyMode.REPLACE_SECTION_A_TO_B) {
+        "Only Replace A-B is supported here."
+    }
+    require(routedPoints.size >= 2) {
+        "The routed replacement is empty."
+    }
+
+    val sourceTrack =
+        GpxTrackDetails(
+            id = sourcePath,
+            points = profile.points.map { it.latLong },
+            title = sourceTitle,
+            distance = profile.totalDistance,
+            elevationGain = profile.totalAscent,
+            startPoint = profile.points.firstOrNull()?.latLong,
+            endPoint = profile.points.lastOrNull()?.latLong,
+        )
+    val a = session.pointA ?: error("Point A is required.")
+    val b = session.pointB ?: error("Point B is required.")
+    val posA = snapToTrackPosition(target = a, sourceTrack = sourceTrack, profile = profile)
+    val posB = snapToTrackPosition(target = b, sourceTrack = sourceTrack, profile = profile)
+    val (start, end) = orderedPositions(posA, posB)
+
+    val merged = ArrayList<TrackPoint>(profile.points.size + routedPoints.size)
+    trimTrackEnd(profile.points, start).forEach { point ->
+        appendUnique(merged, point)
+    }
+    routedPoints.forEach { point ->
+        appendUnique(merged, point.toTrackPoint())
+    }
+    trimTrackStart(profile.points, end).forEach { point ->
+        appendUnique(merged, point)
+    }
+
+    require(merged.size >= 2) {
+        "The selected replacement would produce an empty GPX."
+    }
+
+    val titleBase =
+        sourceTitle?.takeIf { it.isNotBlank() }
+            ?: sourceFileName.removeSuffix(".gpx")
+    val fileName =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> sourceFileName
+            RouteSaveBehavior.SAVE_AS_NEW -> {
+                buildEditedFileName(
+                    sourceFileName = sourceFileName,
+                    mode = session.options.modifyMode,
+                )
+            }
+        }
+    val title =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> titleBase
+            RouteSaveBehavior.SAVE_AS_NEW -> "$titleBase (rerouted)"
+        }
+
+    return RouteToolEditOutput(
+        fileName = fileName,
+        title = title,
+        points = merged,
+    )
+}
+
+internal fun buildRouteToolReshapeOutput(
+    sourcePath: String,
+    sourceFileName: String,
+    sourceTitle: String?,
+    profile: TrackProfile,
+    session: RouteToolSession,
+    firstLegPoints: List<RouteGeometryPoint>,
+    secondLegPoints: List<RouteGeometryPoint>,
+    bounds: RouteReshapeBounds? = null,
+): RouteToolEditOutput {
+    require(session.options.modifyMode == RouteModifyMode.RESHAPE_ROUTE) {
+        "Only reshape edits are supported here."
+    }
+    require(firstLegPoints.size >= 2) {
+        "The first routed reshape leg is empty."
+    }
+    require(secondLegPoints.size >= 2) {
+        "The second routed reshape leg is empty."
+    }
+
+    val selectedPoint = session.pointA ?: error("Select the route point first.")
+    require(session.destination != null) { "Pick the new bend point first." }
+    val resolvedBounds =
+        bounds ?: resolveRouteReshapeBounds(
+            sourcePath = sourcePath,
+            sourceTitle = sourceTitle,
+            profile = profile,
+            anchor = selectedPoint,
+            rejoinHint = session.destination,
+        )
+
+    val merged = ArrayList<TrackPoint>(profile.points.size + firstLegPoints.size + secondLegPoints.size)
+    trimTrackEnd(profile.points, resolvedBounds.startPosition).forEach { point ->
+        appendUnique(merged, point)
+    }
+    firstLegPoints.forEach { point ->
+        appendUnique(merged, point.toTrackPoint())
+    }
+    secondLegPoints.forEach { point ->
+        appendUnique(merged, point.toTrackPoint())
+    }
+    trimTrackStart(profile.points, resolvedBounds.endPosition).forEach { point ->
+        appendUnique(merged, point)
+    }
+
+    require(merged.size >= 2) {
+        "The reshape would produce an empty GPX."
+    }
+
+    val titleBase =
+        sourceTitle?.takeIf { it.isNotBlank() }
+            ?: sourceFileName.removeSuffix(".gpx")
+    val fileName =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> sourceFileName
+            RouteSaveBehavior.SAVE_AS_NEW ->
+                buildEditedFileName(
+                    sourceFileName = sourceFileName,
+                    modeSlug = "reshape",
+                )
+        }
+    val title =
+        when (session.options.saveBehavior) {
+            RouteSaveBehavior.REPLACE_CURRENT -> titleBase
+            RouteSaveBehavior.SAVE_AS_NEW -> "$titleBase (reshaped)"
+        }
+
+    return RouteToolEditOutput(
+        fileName = fileName,
+        title = title,
+        points = merged,
+    )
+}
+
+internal fun buildRouteToolReshapePreview(
+    sourcePath: String,
+    sourceTitle: String?,
+    profile: TrackProfile,
+    session: RouteToolSession,
+    firstLegPoints: List<RouteGeometryPoint>,
+    secondLegPoints: List<RouteGeometryPoint>,
+    bounds: RouteReshapeBounds? = null,
+): RouteToolModifyPreview {
+    require(session.options.modifyMode == RouteModifyMode.RESHAPE_ROUTE) {
+        "Only reshape previews are supported here."
+    }
+    require(firstLegPoints.size >= 2) {
+        "The first routed reshape leg is empty."
+    }
+    require(secondLegPoints.size >= 2) {
+        "The second routed reshape leg is empty."
+    }
+
+    val selectedPoint = session.pointA ?: error("Select the route point first.")
+    require(session.destination != null) { "Pick the new bend point first." }
+    val resolvedBounds =
+        bounds ?: resolveRouteReshapeBounds(
+            sourcePath = sourcePath,
+            sourceTitle = sourceTitle,
+            profile = profile,
+            anchor = selectedPoint,
+            rejoinHint = session.destination,
+        )
+
+    val previewPoints = ArrayList<LatLong>(firstLegPoints.size + secondLegPoints.size + 2)
+    appendUniqueLatLong(previewPoints, resolvedBounds.startPoint.latLong)
+    firstLegPoints.forEach { point ->
+        appendUniqueLatLong(previewPoints, point.latLong)
+    }
+    secondLegPoints.forEach { point ->
+        appendUniqueLatLong(previewPoints, point.latLong)
+    }
+    appendUniqueLatLong(previewPoints, resolvedBounds.endPoint.latLong)
+
+    return RouteToolModifyPreview(previewPoints = previewPoints)
+}
